pyproject.toml
[tool.poetry]
name = "pdf-report-generator"
version = "0.1.0"
description = ""
authors = ["MDiRocco <marc.dirocco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fpdf = "^1.7.2"
matplotlib = "^3.10.0"
fire = "^0.7.0"
pandas = "^2.2.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
